pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraud-ml-platform"
version = "0.1.0"
description = "End-to-end fraud detection ML platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ML Engineering Team"}
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "mlflow>=2.8.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "great-expectations>=0.18.0",
    "evidently>=0.4.0",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
    "kaggle>=1.5.0",
    "requests>=2.31.0",
    "matplotlib>=3.7.0",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
fraud-ingest = "fraud_platform.ingestion.load_raw:main"
fraud-validate = "fraud_platform.validation.run_validation:main"
fraud-features = "fraud_platform.features.build_features:main"
fraud-train = "fraud_platform.training.train:main"
fraud-serve = "fraud_platform.serving.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

